Add support for shapes to opt out of casting shadows
This change adds support to enable / disable for a shape instance to
cast a shadow. This is quite useful in scenarios f.e. when we are
looking through the surface of a pond at some rocks beneath it. In
this case we don't want the surface of the pond to cast any shadows on
the rocks.