Fix asan reported indirect memory leaks
This change fixes the one-and-only asan reported indirect memory leak
in the raytracer.

Leak was happening because child objects in an object-group were
treating their parent objects as shared_ptr. This resulted in a
circular object reference from parent -> child -> parent, causing
'indirect' memory leaks.

The 'fix' was to break this circular reference using the canonical
weak_ptr to store child -> parent references.