Structured rendering 'through' the camera
This is mostly a 'minor' change, revolving around a structured
configuration of various parameters that influence how rendering the
entire scene progresses.

Moreover, scenes can now be rendered in different styles which control
the order in which pixels on the canvas are processed. As of now:

    - scanline and
    - hilbert-curve

orderings are supported.

Apart from this, it seems that with "-ffast-math" option llvm's code
for cube-test fails. Remove that option for now.