Support for show-as-we-go rendering
This is a fairly involved change, and ends up adding support for
'show-as-we-go' i.e. display the contents of the canvas as the
rendering progresses.

Instead of relying on other frameworks f.e. sdl2 etc. this time
around, we are going with X11 using the XCB interface. It is quite
trivial to use !

We can 'visualize' both supported styles of rendering i.e. scanline as
well as hilbert-curve. The latter makes for quite a cool effect,
however, scanline is generally observed to be faster.