Support tile based rendering
This change implements a tile based rendering scheme. As the name
suggests, the entire canvas is divvied up into a grid-of-tiles of
specific dimension. Each of which is then rendered in parallel by as
many hardware threads as are available.

This produces an overall cool looking effect, however it is not as
cool as hilbert style rendering. Also, it is slower than scanline mode
of rendering.

With these 3 styles, we *seem* to be having the following observation:

    - performance: hilbert < tile < scanline
    - coolness:    scanline < tile < hilbert