Move tests closer to the sources which they test
Bunching together all tests in a single top-level 'tests' directory
seems a bit callous. Unit test should be closer to the 'unit' it is
testing.

Towards this end, each of the top-level directory comprising the
ray-tracer now has a 'tests' sub-directory as well, which contains all
the pertinent tests.

Moreover we run the unit tests as part of the build itself. So
breaking changes cannot be made at all.