Support building on MacOS with LLVM and GNU
This is quite a big change and focuses on adding support for building
the ray-tracer on the MacOS platform.

Both LLVM and the GNU toolchain support is added. These toolchains are
obtained via the homebrew package manager.

Newer versions of these toolchains should require just a single line
change, and no code modifications whatsoever.

Moreover, profiling via Tracy is also supported. This should be quite
interesting to experiment on the M4 platform.

A majority of the issue was due to platform specific differences in
low level routines f.e. mmap(2). Which is *not* unexpected. Once these
were fixed, it was pretty much smooth sailing from there onwards.

Minor formatting / beautification changes as well.