Support building on MacOS with LLVM
This is quite a big change and focusses on adding support for building
the ray-tracer on the MacOS platform.

For now, only LLVM support is added. LLVM itself is obtained via
homebrew. Hopefully adding support for GNU toolchain on this platform
should not be too much of a trouble.

It was pretty much a smooth sailing, with some minor annoyances due to
diffferences in system APIs, but that is to be expected.

Minor formatting changes as well.