Add support for Axis Aligned Bounding Boxes
This change adds support for axis-aligned-bounding boxes to
SMRT.

Bounding boxes significantly speed up the raytracer by reducing the
computational load when large groups of objects are involved. This is
done by ensuring that expensive ray-object intersection tests are
performed only when necessary.

For example on the `render_chess_pawn` scene the total rendering time
was reduced by approx. 60% !