Bounding Volume Hierarchy
Add support for BVH in the ray-tracer ! This is a big change, and
has *major* acceleration in rendering scenes with large number of
triangles.

In my cases, I have seen rendering speeds increase by approx. 30x !
Yes, that is really 30x increase over naive rendering.

This just rocks !!!

Ofcourse, as usual tests are also included which exercise this new
interface. Moreover, a trivial program called 'render_dragons' is also
provided which uses this to great effect.